Save the expiration date as a float in the DB
On column declaration the date had a data type of integer and it was
being passed as a String into the insert function. Now it is declared
as a float on both instances

+++ app/src/main/java/com/nicolascarrasco/www/auctionhouse/data/AuctionColumns.java
@@ -28,7 +28,7 @@ public interface AuctionColumns {
     @DataType(DataType.Type.TEXT)
     String DESCRIPTION = "description";
 
-    @DataType(DataType.Type.INTEGER)
+    @DataType(DataType.Type.REAL)
     @DefaultValue("0")
     String PRICE = "price";
